+++ vm_windows.tf
@@ -1,5 +1,5 @@
 resource "azurerm_windows_virtual_machine" "machine" {
-  count                 = local.windows_vm
+  for_each              = var.source_image_reference.offer == "WindowsServer" ? { Windows = "true" } : {}
   name                  = var.name
   resource_group_name   = var.resource_group
   location              = var.location