+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v1.2.0
 - #4 Add the possibility to create a network interface through the module
+- #10 Fix how the module decides what resource to use
 - #11 Remove data.azurerm_resource_group, due to unintentional side-effects
 
 ## v1.1.0